Refactor visualization and cleanup
